refactor: Split public API into multiple files
